WolfAlerts <div> write on alert only, <hr/> fix
When adding a border with the WolfAlerts <div> it was problematic
that the <div> always appeared. Without a border this a non-issue,
but in the end the border was retained and the <div> only appears
on an alert.

Also removed the introduction of an extra <hr/> which I never noticed
above the navigation buttons